--- a/server/src/main/java/org/candlepin/audit/EventBuilder.java
+++ b/server/src/main/java/org/candlepin/audit/EventBuilder.java
@@ -38,7 +38,7 @@ public class EventBuilder {
     private final EventFactory factory;
 
     private Event event;
-    
+
     private static Logger log = LoggerFactory.getLogger(EventBuilder.class);
 
     public EventBuilder(EventFactory factory, Target target, Type type) {