--- a/server/src/main/java/org/candlepin/model/Product.java
+++ b/server/src/main/java/org/candlepin/model/Product.java
@@ -297,7 +297,6 @@ public class Product extends AbstractHibernateObject implements Linkable {
                 .toHashCode();
     }
 
-    
     /**
      * @param content
      */