--- a/server/src/main/java/org/candlepin/sync/CdnImporter.java
+++ b/server/src/main/java/org/candlepin/sync/CdnImporter.java
@@ -65,7 +65,6 @@ public class CdnImporter {
                 //curator.refresh(existing);
                 log.debug("Updating CDN: " + cdn.getName());
             }
-            
         }
     }
 }