--- a/server/src/main/java/org/candlepin/sync/ConsumerImporter.java
+++ b/server/src/main/java/org/candlepin/sync/ConsumerImporter.java
@@ -123,6 +123,7 @@ public class ConsumerImporter {
         owner.setUpstreamConsumer(uc);
 
         curator.merge(owner);
+        //curator.refresh(owner);
     }
 
 }