--- a/server/src/main/java/org/candlepin/sync/ConsumerTypeImporter.java
+++ b/server/src/main/java/org/candlepin/sync/ConsumerTypeImporter.java
@@ -54,7 +54,9 @@ public class ConsumerTypeImporter {
             if (curator.lookupByLabel(consumerType.getLabel()) == null) {
                 curator.create(consumerType);
                 log.debug("Created consumer type: " + consumerType.getLabel());
+                curator.refresh(consumerType);
             }
         }
+        
     }
 }