--- a/server/src/main/java/org/candlepin/sync/ConsumerTypeImporter.java
+++ b/server/src/main/java/org/candlepin/sync/ConsumerTypeImporter.java
@@ -57,6 +57,6 @@ public class ConsumerTypeImporter {
                 curator.refresh(consumerType);
             }
         }
-        
+
     }
 }